+++ packages/vendure-plugin-change-order-customer/src/index.ts
@@ -1,1 +1,0 @@
-export * from './change-order-customer.plugin';