+++ packages/vendure-plugin-popularity-scores/src/popularity-scores.service.ts
@@ -150,7 +150,16 @@ export class PopularityScoresService implements OnModuleInit {
     };
 
     await traverseDepthFirstAndUpdateScore(channelCollectionsTree);
-    await collectionsRepo.save(allCollections);
+    for (const updateCollection of allCollections) {
+      await collectionsRepo.update(
+        { id: updateCollection.id },
+        {
+          customFields: {
+            popularityScore: updateCollection.customFields.popularityScore,
+          },
+        }
+      );
+    }
   }
 
   private async getSummedProductScoreCalculation(