+++ packages/vendure-plugin-popularity-scores/test/dev-server.ts
@@ -14,6 +14,7 @@ import {
 import { PopularityScoresPlugin } from '../src';
 import { initialTestData } from './initial-test-data';
 import { testPaymentMethod } from '../../test/src/test-payment-method';
+import { createSettledOrder } from '../../test/src/shop-utils';
 
 require('dotenv').config();
 
@@ -55,4 +56,9 @@ require('dotenv').config();
 
     productsCsvPath: './test/products.csv',
   });
+  await createSettledOrder(shopClient, 1);
+  await createSettledOrder(shopClient, 1);
+  await createSettledOrder(shopClient, 1);
+  await createSettledOrder(shopClient, 1);
+  await createSettledOrder(shopClient, 1);
 })();