+++ packages/vendure-plugin-primary-collection/src/primary-collection-plugin.ts
@@ -1,16 +1,15 @@
 import {
-  Collection,
   LanguageCode,
   PluginCommonModule,
   RuntimeVendureConfig,
   Type,
   VendurePlugin,
 } from '@vendure/core';
-import { gql } from 'graphql-tag';
-import { PrimaryCollectionResolver } from './api/primary-collection.resolver';
+import { PrimaryCollectionPluginResolver } from './api/primary-collection.resolver';
 import { AdminUiExtension } from '@vendure/ui-devkit/compiler';
 import path from 'path';
 import { PrimaryCollectionHelperService } from './api/primary-collections-helper.service';
+import { apiExtensions } from './api/api-extensions';
 
 export interface PrimaryCollectionPluginConfig {
   /**
@@ -23,23 +22,19 @@ export interface PrimaryCollectionPluginConfig {
   imports: [PluginCommonModule],
   providers: [PrimaryCollectionHelperService],
   shopApiExtensions: {
-    schema: gql`
-      extend type Product {
-        primaryCollection: Collection
-      }
-    `,
-    resolvers: [PrimaryCollectionResolver],
+    schema: apiExtensions,
+    resolvers: [PrimaryCollectionPluginResolver],
+  },
+  adminApiExtensions: {
+    schema: apiExtensions,
+    resolvers: [PrimaryCollectionPluginResolver],
   },
   exports: [PrimaryCollectionHelperService],
   compatibility: '^2.0.0',
   configuration: (config: RuntimeVendureConfig) => {
     config.customFields.Product.push({
       name: 'primaryCollection',
-      type: 'relation',
-      entity: Collection,
-      graphQLType: 'Collection',
-      eager: true,
-      nullable: true,
+      type: 'string',
       ui: {
         component: 'select-primary-collection',
         tab: PrimaryCollectionPlugin?.config?.customFieldUITabName,