+++ packages/vendure-plugin-variant-bulk-update/src/constants.ts
@@ -1,0 +1,2 @@
+export const loggerCtx = 'VariantBulkUpdatePlugin';
+export const PLUGIN_INIT_OPTIONS = Symbol('PLUGIN_INIT_OPTIONS');