+++ packages/vendure-plugin-popularity-scores/CHANGELOG.md
@@ -1,6 +1,10 @@
-# 1.7.1 (2024-06-21)
+# 1.7.2 (2024-07-25)
 
 - Allow specifying a UI tab for the `popularityScore` custom field
+
+# 1.7.1 (2024-07-22)
+
+- Fixed issue where in some projects a collection's parentId was set to null. See #478
 
 # 1.7.0 (2024-06-21)